+++ internal/gitops/structure/common/types.go
@@ -3,6 +3,8 @@ package common
 import (
 	"os"
 
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	"github.com/giantswarm/kubectl-gs/internal/gitops/encryption"
 )
 
@@ -10,8 +12,12 @@ type StructureConfig struct {
 	App                    string
 	AppBase                string
 	AppCatalog             string
+	AppInstallTimeout      *metav1.Duration
 	AppName                string
 	AppNamespace           string
+	AppRollbackTimeout     *metav1.Duration
+	AppUninstallTimeout    *metav1.Duration
+	AppUpgradeTimeout      *metav1.Duration
 	AppUserValuesConfigMap string
 	AppUserValuesSecret    string
 	AppVersion             string