+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Added support for deriving internal API hostname from workload cluster main API URLs
+
 ## [2.23.1] - 2022-09-27
 
 ### Fixed