+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
+- Add timeouts support to App CR.
 - Added support for deriving internal API hostname from workload cluster main API URLs
 
 ## [2.23.1] - 2022-09-27