+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [2.23.2] - 2022-10-04
+
 ### Added
 
 - Add timeouts support to App CR.
@@ -1037,7 +1039,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v2.23.1...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v2.23.2...HEAD
+[2.23.2]: https://github.com/giantswarm/kubectl-gs/compare/v2.23.1...v2.23.2
 [2.23.1]: https://github.com/giantswarm/kubectl-gs/compare/v2.23.0...v2.23.1
 [2.23.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.22.0...v2.23.0
 [2.22.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.21.0...v2.22.0